+++ mycodo/databases/models/misc.py
@@ -1,8 +1,9 @@
 # coding=utf-8
 import logging
 
-from mycodo.databases import CRUDMixin
-from mycodo.databases import set_uuid
+import requests
+
+from mycodo.databases import CRUDMixin, set_uuid
 from mycodo.mycodo_flask.extensions import db
 
 logger = logging.getLogger("mycodo.misc")
@@ -48,21 +49,32 @@ class Misc(CRUDMixin, db.Model):
     default_login_page = db.Column(db.String, default='password')
 
     # Measurement database
-    measurement_db_name = db.Column(db.String, default='influxdb')
-    measurement_db_version = db.Column(db.String, default='1')
+    influx_db_version = '1'
+    influxdb_host = 'localhost'
+
+    try:
+        from mycodo.scripts.measurement_db import get_influxdb_host
+        tmp_host = get_influxdb_host()
+        if tmp_host:
+            influxdb_host = tmp_host
+    except:
+        logger.exception("Determining influxdb host")
+
     try:
         from mycodo.scripts.measurement_db import get_influxdb_info
         influx_info = get_influxdb_info()
         if influx_info['influxdb_installed']:
             if influx_info['influxdb_version'].startswith('1'):
-                measurement_db_version = db.Column(db.String, default='1')
+                influx_db_version = '1'
             elif influx_info['influxdb_version'].startswith('2'):
-                measurement_db_version = db.Column(db.String, default='2')
+                influx_db_version = '2'
     except:
-        logger.exception("creating influxdb options")
+        logger.exception("Determining influxdb version")
 
-    measurement_db_host = db.Column(db.String, default='localhost')
+    measurement_db_version = db.Column(db.String, default=influx_db_version)
+    measurement_db_host = db.Column(db.String, default=influxdb_host)
     measurement_db_port = db.Column(db.String, default=8086)
+    measurement_db_name = db.Column(db.String, default='influxdb')
     measurement_db_user = db.Column(db.String, default='mycodo')
     measurement_db_password = db.Column(db.String, default='mmdu77sj3nIoiajjs')
     measurement_db_dbname = db.Column(db.String, default='mycodo_db')