+++ CHANGELOG.md
@@ -56,6 +56,8 @@ New Behavior: InfluxDB 1.x returned data as an epoch float value (e.g. 166722857
  - Fix MCP23017 Pump Output initialization
  - Fix race condition with Kasa RGB Bulb that could sometimes turn it back on after turning off
  - Fix digits not properly displaying on PID Controller Widget
+ - Fix determining improper influxdb host in Docker
+ - Fix flask and nginx logs appearing empty in Docker
 
 ### Miscellaneous